feito ajuste na logica para excluir usuario

+++ stool-in-app-ui/lib/features/auth/presenter/sign_in/cubit/sign_in_cubit.dart
@@ -114,7 +114,9 @@ class SignInCubit extends Cubit<SignInState> with SharedPreferencesHelper {
     );
     result.fold(
       (error) async {
-        await _firebaseAuth.currentUser?.delete();
+        if (userUuid == null) {
+          await _firebaseAuth.currentUser?.delete();
+        }
         emit(
           SignInStateError(message: error.message),
         );
